requirements.txt
celery>=5.3.0
redis>=5.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
webdriver-manager>=4.0.0
google-generativeai>=0.3.0
lxml>=4.9.0